backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1

# Google Gemini SDK
google-generativeai==0.8.3

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.11.5
python-docx==1.1.2
Pillow==11.1.0
PyYAML==6.0.2

# Data Validation
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Database & Cache
psycopg2-binary==2.9.10
redis==7.1.0
sqlalchemy==2.0.37

# Async & Background Tasks
celery==5.4.0
aiofiles==24.1.0

# Security & Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
httpx[cli]==0.28.1
